Rename get + set container settings to get + set instance settings
