+++ src/classes/Tools/Instances/SetInstanceSettings.php
@@ -1,10 +1,10 @@
 <?php
 
-namespace dhope0000\LXDClient\Tools\Containers;
+namespace dhope0000\LXDClient\Tools\Instances;
 
 use dhope0000\LXDClient\Model\Client\LxdClient;
 
-class SetContainerSettings
+class SetInstanceSettings
 {
     public function __construct(LxdClient $lxdClient)
     {
